% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awntool-package.R
\docType{package}
\name{awntool-package}
\alias{awntool}
\alias{awntool-package}
\title{awntool: Shiny App for querying official apartment numbers in the city of Zurich}
\description{
Shiny App aWN-Tool that shows official aparment numbers for all the addresses in the city of Zurich.
}
\author{
\strong{Maintainer}: Alessandro Feller \email{alessandro.feller@zuerich.ch}

Authors:
\itemize{
  \item Thamilini Easwaralingam \email{thamilini.easwaralingam@zuerich.ch}
}

Other contributors:
\itemize{
  \item Statistik Stadt Zürich [funder]
}

}
\keyword{internal}
