% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awntool-package.R
\docType{package}
\name{awntool-package}
\alias{awntool}
\alias{awntool-package}
\title{awntool: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
