% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awntool-package.R
\docType{package}
\name{aWNtool-package}
\alias{aWNtool}
\alias{aWNtool-package}
\title{aWNtool: Shiny App for querying official apartment numbers in the city of Zurich}
\description{
Shiny App aWN-Tool that shows official aparment numbers for all the addresses in the city of Zurich.
}
\author{
\strong{Maintainer}: Alessandro Feller \email{alessandro.feller@zuerich.ch}

Authors:
\itemize{
  \item Thamilini Easwaralingam \email{thamilini.easwaralingam@zuerich.ch}
}

}
\keyword{internal}
