% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awntool-package.R
\docType{package}
\name{aWNtool-package}
\alias{aWNtool}
\alias{aWNtool-package}
\title{aWNtool: Shiny App for querying official apartment numbers in the city of Zurich}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thamilini Easwaralingam \email{thamilini.easwaralingam@zuerich.ch}

}
\keyword{internal}
